requirement.txt
RPi.GPIO
qrcode
